pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bigfoont"
version = "0.1.0"
description = "Render text as braille unicode art using bitmap fonts"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pillow>=10.0.0",
]

[project.scripts]
bigfoont = "bigfoont.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
include = ["bigfoont*"]
